move class to new file
