+++ connector/src/yang/connector/grpc/__init__.py
@@ -1,0 +1,78 @@
+import logging
+import tempfile
+from importlib import import_module
+from shutil import copyfile
+
+from pyats.connections import BaseConnection
+from pyats.easypy import runtime
+
+# create a logger for this module
+log = logging.getLogger(__name__)
+
+
+class Grpc(BaseConnection):
+    """Session handling for Grpc outbound connections.
+
+        Can be used with pyATS same as yang.connector
+
+        EXAMPLE USAGE
+
+        Welcome to pyATS Interactive Shell
+        ==================================
+        Python 3.11.5 (main, Sep 25 2023, 16:57:00) [Clang 14.0.0 (clang-1400.0.29.202)]
+
+        >>> from pyats.topology.loader import load
+        >>> testbed = load('/Users/user/testbed.yaml')
+        -------------------------------------------------------------------------------
+        >>> dev = testbed.devices['router-1']
+        >>> dev.connect(via='grpc', alias='grpc')
+
+    """
+    def __new__(cls, *args, **kwargs):
+        if '.'.join([cls.__module__, cls.__name__]) == \
+                'yang.connector.grpc.Grpc':
+            transporter = kwargs.get('transporter', 'telegraf')
+            mod = import_module(f'yang.connector.grpc.{transporter}')
+            new_cls = mod.Grpc
+            return super().__new__(new_cls)
+        return super().__new__(cls)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.device = kwargs.get('device')
+        dev_args = self.connection_info
+        self.log = log
+        self.log.setLevel(logging.INFO)
+
+        protocol = dev_args.get('protocol', 'grpc').lower()
+        if protocol != 'grpc':
+            msg = f"Invalid protocol {protocol}"
+            raise TypeError(msg)
+
+        self.username = dev_args.get('username', '')
+        self.password = dev_args.get('password', '')
+        if not self.username or not self.password:
+            credentials = dev_args.get('credentials', '')
+            if not credentials:
+                raise KeyError("No credentials found for testbed")
+            if 'grpc' not in credentials:
+                log.info(f"Credentials used from {next(iter(credentials))}")
+            grpc_uname_pwd = credentials.get('')
+            if not grpc_uname_pwd:
+                raise KeyError('No credentials found for gRPC testbed')
+
+        self.host = dev_args.get('host') or dev_args.get('ip')
+
+        self.overwrite = dev_args.get('overwrite_config_file', False)
+        if self.overwrite:
+            self.config_directory = runtime.directory
+        else:
+            self.config_directory = tempfile.mkdtemp()
+        self.output_file = dev_args.get('output_file', f"{runtime.directory}/mdt")
+        try:
+            self.config_file = copyfile(dev_args.get('config_file', None), f"{runtime.directory}/telegraf.conf")
+        except TypeError:
+            self.config_file = None
+        self.telemetry_subscription_id = dev_args.get('telemetry_subscription_id', 11172017)
+
+        self.transport_process = None
