+++ connector/src/yang/connector/grpc/telegraf.py
@@ -1,102 +1,21 @@
 import logging
-import multiprocessing
-import re
 import socket
 import configparser
 import subprocess
 import tempfile
-from shutil import copyfile
 
 from pyats.connections import BaseConnection
 from pyats.easypy import runtime
 from pyats.utils.secret_strings import to_plaintext
-from pyats.easypy import runtime
-from genie.libs.sdk.apis.utils import get_local_ip
-from genie.libs.sdk.apis.iosxe.telemetry.configure import *
 from unicon import Connection
 
+from . import Grpc
 
 # create a logger for this module
 log = logging.getLogger(__name__)
 
 
-class Grpc(BaseConnection):
-    """Session handling for Grpc outbound connections.
-
-        Can be used with pyATS same as yang.connector
-
-        EXAMPLE USAGE
-
-        Welcome to pyATS Interactive Shell
-        ==================================
-        Python 3.11.5 (main, Sep 25 2023, 16:57:00) [Clang 14.0.0 (clang-1400.0.29.202)]
-
-        >>> from pyats.topology.loader import load
-        >>> testbed = load('/Users/user/testbed.yaml')
-        -------------------------------------------------------------------------------
-        >>> dev = testbed.devices['router-1']
-        >>> dev.connect(via='grpc', alias='grpc')
-
-    """
-    def __new__(cls, *args, **kwargs):
-        transporter = kwargs.get('transporter', 'telegraf')
-        if transporter.lower() == 'telegraf':
-            return super().__new__(GrpcTelegraf)
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.device = kwargs.get('device')
-        dev_args = self.connection_info
-        self.log = log
-        self.log.setLevel(logging.INFO)
-
-        protocol = dev_args.get('protocol', 'grpc').lower()
-        if protocol != 'grpc':
-            msg = f"Invalid protocol {protocol}"
-            raise TypeError(msg)
-
-        self.username = dev_args.get('username', '')
-        self.password = dev_args.get('password', '')
-        if not self.username or not self.password:
-            credentials = dev_args.get('credentials', '')
-            if not credentials:
-                raise KeyError("No credentials found for testbed")
-            if 'grpc' not in credentials:
-                log.info(f"Credentials used from {next(iter(credentials))}")
-            grpc_uname_pwd = credentials.get('')
-            if not grpc_uname_pwd:
-                raise KeyError('No credentials found for gRPC testbed')
-
-        self.host = dev_args.get('host') or dev_args.get('ip')
-
-        self.overwrite = dev_args.get('overwrite_config_file', False)
-        if self.overwrite:
-            self.config_directory = runtime.directory
-        else:
-            self.config_directory = tempfile.mkdtemp()
-        self.transporter = dev_args.get('transporter', 'telegraf').lower()
-        try:
-            self.output_file = copyfile(dev_args.get('output_file', None), f"{runtime.directory}/mdt")
-        except TypeError:
-            self.output_file = f"{runtime.directory}/mdt"
-        try:
-            self.config_file = copyfile(dev_args.get('config_file', None), f"{runtime.directory}/telegraf.conf")
-        except TypeError:
-            self.config_file = None
-        self.telemetry_subscription_id = dev_args.get('telemetry_subscription_id', 11172017)
-
-        self.transport_process = None
-
-    def connect(self):
-        raise NotImplementedError
-
-    def disconnect(self):
-        self.transport_process.terminate()
-        self.device.api.unconfigure_telemetry_ietf_subscription(self.telemetry_subscription_id)
-        self.device.disconnect()
-
-
-class GrpcTelegraf(Grpc):
+class Grpc(Grpc):
     """
     Subclass for using Telegraf to collect telemetry
     """
@@ -191,3 +110,8 @@ class GrpcTelegraf(Grpc):
         self.device.api.configure_telemetry_ietf_parameters(self.telemetry_subscription_id,
                                                             "yang-push", local_ip, allocated_port, "grpc-tcp")
         log.info(f"Started gRPC inbound server on {local_ip}:{allocated_port}")
+
+    def disconnect(self):
+        self.transport_process.terminate()
+        self.device.api.unconfigure_telemetry_ietf_subscription(self.telemetry_subscription_id)
+        self.device.disconnect()
